requirements.txt
tqdm
yara-python
python-dotenv
filetype